Move helper functions to new header/impl file pair
