+++ src/scripting/CMakeLists.txt
@@ -225,6 +225,8 @@ wl_library(scripting_logic
     ui/lua_textarea.h
     ui/lua_text_input_panel.cc
     ui/lua_text_input_panel.h
+    ui/lua_ui_helper_functions.cc
+    ui/lua_ui_helper_functions.h
     ui/lua_window.cc
     ui/lua_window.h
   USES_SDL2
